feat(project): Install Vuex and split TaskList
