+++ taskbox/package.json
@@ -3,16 +3,12 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "test:unit": "vue-cli-service test:unit",
-    "lint": "vue-cli-service lint",
-    "storybook": "start-storybook -p 6006",
-    "build-storybook": "build-storybook"
+    "storybook": "start-storybook -p 6006 -s public"
   },
   "dependencies": {
     "core-js": "^3.6.5",
-    "vue": "^2.6.11"
+    "vue": "^2.6.11",
+    "vuex": "^3.5.1"
   },
   "devDependencies": {
     "@babel/core": "^7.12.3",
@@ -34,8 +30,5 @@
     "prettier": "^1.19.1",
     "react-is": "^17.0.1",
     "vue-template-compiler": "^2.6.11"
-  },
-  "scripts": {
-    "storybook": "start-storybook -p 6006 -s public"
   }
 }
