+++ taskbox/src/App.vue
@@ -1,30 +1,23 @@
+<!--src/App.vue-->
 <template>
   <div id="app">
-    <img alt="Vue logo" src="./assets/logo.png" />
-    <HelloWorld msg="Welcome to Your Vue.js App" />
+    <task-list />
   </div>
 </template>
 
 <script>
-import HelloWorld from "./components/HelloWorld.vue";
+  import store from './store';
+  import TaskList from './components/TaskList.vue';
 
-export default {
-  name: "App",
-  components: {
-    HelloWorld
-  }
-};
+  export default {
+    name: 'app',
+    store,
+    components: {
+      TaskList,
+    },
+  };
 </script>
 
 <style>
   @import './index.css';
-
-  #app {
-    font-family: Avenir, Helvetica, Arial, sans-serif;
-    -webkit-font-smoothing: antialiased;
-    -moz-osx-font-smoothing: grayscale;
-    text-align: center;
-    color: #2c3e50;
-    margin-top: 60px;
-  }
 </style>
