+++ taskbox/src/components/TaskList.vue
@@ -1,61 +1,24 @@
-<!--src/components/TaskList.vue-->
+<!--src/components/TaskList.vue`-->
 <template>
   <div>
-    <div v-if="loading">
-      <div class="loading-item" v-for="(n, index) in 5" :key="index">
-        <span class="glow-checkbox" />
-        <span class="glow-text"> <span>Loading</span> <span>cool</span> <span>state</span> </span>
-      </div>
-    </div>
-    <div class="list-items" v-if="noTasks && !this.loading">
-      <div class="wrapper-message">
-        <span class="icon-check" />
-        <div class="title-message">You have no tasks</div>
-        <div class="subtitle-message">Sit back and relax</div>
-      </div>
-    </div>
-    <div class="list-items" v-if="showTasks">
-      <task
-        v-for="(task, index) in tasksInOrder"
-        :key="index"
-        :task="task"
-        @archiveTask="$emit('archiveTask', $event)"
-        @pinTask="$emit('pinTask', $event)"
-      />
-    </div>
+    <pure-task-list :tasks="tasks" @archiveTask="archiveTask" @pinTask="pinTask" />
   </div>
 </template>
 
 <script>
-  import Task from './Task';
+  import PureTaskList from './PureTaskList';
+  import { mapState, mapActions } from 'vuex';
+
   export default {
     name: 'task-list',
-    props: {
-      loading: {
-        type: Boolean,
-        default: false,
-      },
-      tasks: {
-        type: Array,
-        default: () => []
-      },
-    },
     components: {
-      Task,
+      PureTaskList,
+    },
+    methods: {
+      ...mapActions(['archiveTask', 'pinTask']),
     },
     computed: {
-      noTasks() {
-        return this.tasks.length === 0;
-      },
-      showTasks() {
-        return !this.loading && !this.noTasks;
-      },
-      tasksInOrder() {
-        return [
-          ...this.tasks.filter(t => t.state === 'TASK_PINNED'),
-          ...this.tasks.filter(t => t.state !== 'TASK_PINNED'),
-        ];
-      },
+      ...mapState(['tasks']),
     },
   };
 </script>
