+++ taskbox/tests/unit/TaskList.spec.js
@@ -1,15 +1,15 @@
 //tests/unit/TaskList.spec.js
 import Vue from 'vue';
-import TaskList from '../../src/components/TaskList.vue';
-import { withPinnedTasksData } from '../../src/components/TaskList.stories';
+import PureTaskList from '../../src/components/PureTaskList.vue';
+import { withPinnedTasksData } from '../../src/components/PureTaskList.stories';
 
 it('renders pinned tasks at the start of the list', () => {
-  const Constructor = Vue.extend(TaskList);
+  const Constructor = Vue.extend(PureTaskList);
   const vm = new Constructor({
     propsData: { tasks: withPinnedTasksData },
   }).$mount();
-  const firstTaskPinned = vm.$el.querySelector('.list-item:nth-child(1).TASK_PINNED');
+  const lastTaskInput = vm.$el.querySelector('.list-item:nth-child(1).TASK_PINNED');
 
   // We expect the pinned task to be rendered first, not at the end
-  expect(firstTaskPinned).not.toBe(null);
+  expect(lastTaskInput).not.toBe(null);
 });
